Add a fastapi lifespan with db connection test. Update makefile and dockerfile with fixes
