+++ apps/api/Dockerfile.api
@@ -10,4 +10,4 @@ RUN uv sync
 
 EXPOSE 80
 
-CMD ["uv", "run", "uvicorn", "server:app", "--host", "0.0.0.0", "--port", "80"]
+CMD ["uv", "run", "uvicorn", "src.server:app", "--host", "0.0.0.0", "--port", "80"]
