+++ apps/api/src/__init__.py
@@ -1,0 +1,3 @@
+from logging import getLogger
+
+logger = getLogger(__name__)
